+++ weldx/visualization/k3d_impl.py
@@ -289,6 +289,8 @@ class SpatialDataVisualizer:
         if isinstance(data, geo.SpatialData):
             triangles = data.triangles
             data = data.coordinates.data
+        # k3d needs single precision data.
+        data = data.astype(np.float32)
 
         self._reference_system = reference_system
 
@@ -427,7 +429,7 @@ class CoordinateSystemManagerVisualizerK3D:
         colors :
             A mapping between a coordinate system name or a data set name and a color.
             The colors must be provided as 24 bit integer values that are divided into
-            three 8 bit sections for the rgb values. For example `0xFF0000` for pure
+            three 8 bit sections for the rgb values. For example ``0xFF0000`` for pure
             red.
             Each coordinate system or data set that does not have a mapping in this
             dictionary will get a default color assigned to it.